blueprints/data-solutions/data-platform-minimal/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-cloud-storage>=3.1.1",
    "pandas>=2.3.0",
    "pyarrow>=20.0.0",
]
